# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen rental PlayStation Byone Arena. Mendukung realtime via WebSocket untuk aplikasi mobile dan TV Android.",
        "title": "Byone Arena API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Byone Arena Support",
            "email": "support@byone-arena.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/food-orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan seluruh pesanan makanan diurutkan terbaru",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Ambil semua pesanan makanan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.FoodOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin membuat pesanan makanan. Harga otomatis diambil dari data menu, total dihitung otomatis dari quantity × harga.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Buat pesanan makanan baru",
                "parameters": [
                    {
                        "description": "Data pesanan baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateFoodOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.FoodOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/food-orders/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter pesanan berdasarkan status: pending, preparing, served, cancelled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Ambil pesanan berdasarkan status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status pesanan",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.FoodOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/food-orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan detail pesanan beserta semua item dan info menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Ambil pesanan berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food Order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.FoodOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus pesanan secara permanen. Hanya pesanan berstatus 'cancelled' yang bisa dihapus.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Hapus pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food Order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/food-orders/{id}/cancel": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membatalkan pesanan makanan. Tidak bisa membatalkan pesanan yang sudah served.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Batalkan pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food Order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.FoodOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/food-orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah status pesanan. Alur: pending → preparing → served. Dapat dibatalkan (cancelled) dari status apapun kecuali served.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Update status pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food Order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.FoodOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{session_id}/food-orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan semua pesanan makanan yang terhubung ke sesi PS tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FoodOrder"
                ],
                "summary": "Ambil pesanan makanan dari satu sesi PS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (UUID)",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.FoodOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isMember": {
                    "description": "pelanggan member mendapat diskon otomatis",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "entity.FoodOrder": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/entity.Customer"
                },
                "customerId": {
                    "description": "opsional: walk-in",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FoodOrderItem"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "string"
                },
                "session": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Session"
                        }
                    ]
                },
                "sessionId": {
                    "description": "opsional: terhubung ke sesi PS",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.OrderStatus"
                },
                "totalAmount": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.FoodOrderItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "menuItem": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.MenuItem"
                        }
                    ]
                },
                "menuItemId": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "orderId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "subtotal": {
                    "description": "quantity × unit_price",
                    "type": "number"
                },
                "unitPrice": {
                    "description": "snapshot harga saat pesan",
                    "type": "number"
                }
            }
        },
        "entity.MenuItem": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entity.MenuItemCategory"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "entity.OrderStatus": {
            "type": "string",
            "enum": [
                "pending",
                "preparing",
                "served",
                "cancelled"
            ],
            "x-enum-comments": {
                "OrderStatusCancelled": "dibatalkan",
                "OrderStatusPending": "pesanan baru, menunggu diproses dapur",
                "OrderStatusPreparing": "sedang diproses dapur",
                "OrderStatusServed": "sudah diantarkan ke pelanggan"
            },
            "x-enum-varnames": [
                "OrderStatusPending",
                "OrderStatusPreparing",
                "OrderStatusServed",
                "OrderStatusCancelled"
            ]
        },
...
        "entity.Session": {
            "type": "object",
            "properties": {
                "bookedDurationMinutes": {
                    "description": "BookedDurationMinutes adalah durasi yang dipesan di awal (contoh: 120 menit = 2 jam)",
                    "type": "integer"
                },
                "console": {
                    "description": "Relasi (join data)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Console"
                        }
                    ]
                },
                "consoleId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/entity.Customer"
                },
                "customerId": {
                    "description": "nullable, walk-in",
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "endScheduledAt": {
                    "description": "EndScheduledAt adalah waktu selesai yang direncanakan (StartTime + BookedDurationMinutes)",
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.SessionStatus"
                },
                "totalPrice": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "usecase.CreateFoodOrderRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "customerId": {
                    "description": "opsional: pelanggan terdaftar",
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/usecase.FoodOrderItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "sessionId": {
                    "description": "opsional: sesi PS yang sedang berjalan",
                    "type": "string"
                }
            }
        },
...
        "usecase.FoodOrderItemRequest": {
            "type": "object",
            "required": [
                "menuItemId",
                "quantity"
            ],
            "properties": {
                "menuItemId": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "usecase.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "preparing",
                        "served",
                        "cancelled"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}